requirements.txt
# Avalanche App - Core Dependencies
# Install with: pip install -r requirements.txt

# Web framework
streamlit>=1.50.0
streamlit-authenticator>=0.4.0
streamlit-folium>=0.20.0

# Geospatial
geopandas>=1.0.0
rasterio>=1.4.0
pyproj>=3.6.0
shapely>=2.0.0
fiona>=1.10.0

# Data processing
numpy>=1.26.0
pandas>=2.0.0
scipy>=1.12.0

# Visualization
matplotlib>=3.8.0
plotly>=6.0.0
folium>=0.18.0

# Background task processing
celery>=5.4.0
redis>=5.0.0

# Authentication
bcrypt>=4.0.0
PyJWT>=2.8.0

# Utilities
PyYAML>=6.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
psutil>=6.0.0

# AvaFrame (install separately)
# cd AvaFrame && pip install -e .
